reimplement SA00056
SA00095 test case was detected as wrong, because an additonal start
receive was inserted. Set createInstance to no does not change the
effect thats tested in SA00095, but let the test bas SA00056.

Three further tests are now violating SA00056. The problem is the same
for all three tests: The process is started in the flow. Putting the
start receive outside of the flow does not change the tested effect. but
this must be done in the betsy tests.